backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.12
pillow>=11.0.0
transformers>=4.45.0
torch>=2.4,<3
numpy<2
huggingface-hub>=0.24.0
python-dotenv>=1.0.0
